Add update ListView and DetailView extracting async calls to api module
