+++ packages/transactions/src/api.ts
@@ -1,4 +1,7 @@
-import { gql } from 'graphql-request';
+import { Environment } from './../../config/src/index';
+import { QueryProps } from './../../common/src/ui/layout/tables/types';
+import { Transaction } from './../../common/src/types/index';
+import { request, gql } from 'graphql-request';
 
 export const getDetailQuery = (): string => gql`
   query transaction($id: String!) {
@@ -46,3 +49,41 @@ export const getListQuery = (): string => gql`
     }
   }
 `;
+
+export const deleteFn = async (transactions: Transaction[]) => {
+  await request(Environment.API_URL, getDeleteMutation(), {
+    transactions,
+  });
+};
+
+export const listQueryFn = async (
+  queryParams: QueryProps<Transaction>
+): Promise<{ data: Transaction[]; totalLength: number }> => {
+  const { first, offset, sortBy } = queryParams;
+
+  const { transactions } = await request(Environment.API_URL, getListQuery(), {
+    first,
+    offset,
+    sort: sortBy?.[0]?.id,
+    desc: !!sortBy?.[0]?.desc,
+  });
+
+  return transactions;
+};
+
+export const detailQueryFn = (id: string) => async (): Promise<Transaction> => {
+  const result = await request(Environment.API_URL, getDetailQuery(), {
+    id,
+  });
+  const { transaction } = result;
+  return transaction;
+};
+
+export const updateTransactionFn = async (
+  updated: Transaction
+): Promise<Transaction> => {
+  const patch = { transactionPatch: updated };
+  const result = await request(Environment.API_URL, getMutation(), patch);
+  const { upsertTransaction } = result;
+  return upsertTransaction;
+};
